requirements.txt
amqp==2.5.1
billiard==3.5.0.5
boto==2.48.0
boto3==1.5.4
botocore==1.8.50
celery==4.2.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-filter==2.2.0
django-import-export==1.2.0
django-storages==1.6.5
django-timezone-field==3.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
et-xmlfile==1.0.1
gunicorn==19.7.1
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
jmespath==0.9.4
kombu==4.6.5
MarkupPy==1.14
more-itertools==7.2.0
odfpy==1.4.0
openpyxl==3.0.0
Pillow==5.2.0
postmarker==0.13.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
PySocks==1.7.1
python-crontab==2.4.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.5
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.1.13
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
twilio==6.25.1
urllib3==1.25.6
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
